+++ simulator/runtime/calc/operation/calc_unary.cpp
@@ -19,8 +19,22 @@
 OPEN_RDO_RUNTIME_NAMESPACE
 
 // --------------------------------------------------------------------------------
+// -------------------- RDOCalcUnaryBase
+// --------------------------------------------------------------------------------
+RDOCalcUnaryBase::RDOCalcUnaryBase(const LPRDOCalc& pOperation)
+	: m_pOperation(pOperation)
+{
+	ASSERT(m_pOperation);
+}
+
+// --------------------------------------------------------------------------------
 // -------------------- Унарные операции
 // --------------------------------------------------------------------------------
+RDOCalcDoubleToIntByResult::RDOCalcDoubleToIntByResult(const LPRDOCalc& pOper)
+	: m_round     (false)
+	, m_pOperation(pOper)
+{}
+
 RDOValue RDOCalcDoubleToIntByResult::doCalc(const LPRDORuntime& pRuntime)
 {
 	RDOValue value = m_pOperation->calcValue(pRuntime);
@@ -29,9 +43,18 @@ RDOValue RDOCalcDoubleToIntByResult::doCalc(const LPRDORuntime& pRuntime)
 		: value;
 }
 
+void RDOCalcDoubleToIntByResult::needRound()
+{
+	m_round = true;
+}
+
 // --------------------------------------------------------------------------------
 // -------------------- RDOCalcInt (приведение к целому)
 // --------------------------------------------------------------------------------
+RDOCalcInt::RDOCalcInt(const LPRDOCalc& pOperation)
+	: m_pOperation(pOperation)
+{}
+
 RDOValue RDOCalcInt::doCalc(const LPRDORuntime& pRuntime)
 {
 	RDOValue res = m_pOperation->calcValue(pRuntime);
