--- simulator/runtime/process/advance-inl.h
@@ -1,33 +1,0 @@
-/*!
-  \copyright (c) RDO-Team, 2012
-  \file      advance-inl.h
-  \authors   Урусов Андрей (rdo@rk9.bmstu.ru)
-  \authors   Лущан Дмитрий (dluschan@rk9.bmstu.ru)
-  \date      12.02.2012
-  \brief     Процессный оператор ADVANCE
-  \indent    4T
-*/
-
-// ----------------------------------------------------------------------- INCLUDES
-// ----------------------------------------------------------------------- SYNOPSIS
-#include "simulator/runtime/rdo.h"
-#include "simulator/runtime/rdo_runtime.h"
-#include "simulator/runtime/process/rdoprocess_i.h"
-// --------------------------------------------------------------------------------
-
-OPEN_RDO_RUNTIME_NAMESPACE
-
-// --------------------------------------------------------------------------------
-// -------------------- RDOPROCAdvance
-// --------------------------------------------------------------------------------
-inline RDOPROCAdvance::RDOPROCAdvance(LPIPROCProcess process, const LPRDOCalc& _pDelayCalc)
-	: RDOPROCBlock(process    )
-	, pDelayCalc  (_pDelayCalc)
-{}
-
-inline RDOPROCAdvance::LeaveTr::LeaveTr(const LPRDOPROCTransact& _transact, double _timeLeave)
-	: transact (_transact )
-	, timeLeave(_timeLeave)
-{}
-
-CLOSE_RDO_RUNTIME_NAMESPACE
