+++ simulator/runtime/process/advance.cpp
@@ -23,6 +23,16 @@ OPEN_RDO_RUNTIME_NAMESPACE
 // --------------------------------------------------------------------------------
 // -------------------- RDOPROCAdvance
 // --------------------------------------------------------------------------------
+RDOPROCAdvance::RDOPROCAdvance(LPIPROCProcess process, const LPRDOCalc& _pDelayCalc)
+	: RDOPROCBlock(process    )
+	, pDelayCalc  (_pDelayCalc)
+{}
+
+RDOPROCAdvance::LeaveTr::LeaveTr(const LPRDOPROCTransact& _transact, double _timeLeave)
+	: transact (_transact )
+	, timeLeave(_timeLeave)
+{}
+
 bool RDOPROCAdvance::onCheckCondition(const LPRDORuntime& pRuntime)
 {
 	if (!m_transacts.empty())
