+++ simulator/runtime/process/assign.h
@@ -47,6 +47,4 @@ private:
 
 CLOSE_RDO_RUNTIME_NAMESPACE
 
-#include "simulator/runtime/process/assign-inl.h"
-
 #endif // _LIB_RUNTIME_PROCESS_ASSIGN_H_
