+++ simulator/runtime/process/generate.h
@@ -89,6 +89,4 @@ private:
 
 CLOSE_RDO_RUNTIME_NAMESPACE
 
-#include "simulator/runtime/process/generate-inl.h"
-
 #endif // _LIB_RUNTIME_PROCESS_GENERATE_H_
