+++ simulator/runtime/process/rdoprocess.cpp
@@ -21,6 +21,50 @@
 OPEN_RDO_RUNTIME_NAMESPACE
 
 // --------------------------------------------------------------------------------
+// -------------------- RDOPROCBlock
+// --------------------------------------------------------------------------------
+RDOPROCBlock::RDOPROCBlock(LPIPROCProcess pProcess)
+	: m_process(pProcess)
+{}
+
+RDOPROCBlock::~RDOPROCBlock()
+{}
+
+bool RDOPROCBlock::init()
+{
+	if (!m_process)
+		return false;
+
+	m_process.query_cast<IBaseOperationContainer>()->append(this);
+	return true;
+}
+
+RDOPROCBlock::TransactIt RDOPROCBlock::transactFind(const LPTransact& pTransact)
+{
+	return std::find(m_transacts.begin(), m_transacts.end(), pTransact);
+}
+
+RDOPROCBlock::TransactIt RDOPROCBlock::transactEnd()
+{
+	return m_transacts.end();
+}
+
+void RDOPROCBlock::transactGoIn(const LPTransact& pTransact)
+{
+	m_transacts.push_back(pTransact);
+}
+
+void RDOPROCBlock::transactGoOut(const LPTransact& pTransact)
+{
+	m_transacts.remove(pTransact);
+}
+
+LPIPROCProcess RDOPROCBlock::getProcess() const
+{
+	return m_process;
+}
+
+// --------------------------------------------------------------------------------
 // -------------------- RDOPROCProcess
 // --------------------------------------------------------------------------------
 RDOPROCProcess::RDOPROCProcess(const std::string& name, const LPRDORuntime& pRuntime)
@@ -116,6 +160,26 @@ RDOPROCTransact::RDOPROCTransact(const LPRDORuntime& pRuntime, const std::vector
 RDOPROCTransact::~RDOPROCTransact()
 {}
 
+LPRDOPROCResource RDOPROCTransact::getRes()
+{
+	return m_res;
+}
+
+void RDOPROCTransact::setRes(const LPRDOPROCResource& pResource)
+{
+	m_res = pResource;
+}
+
+LPIPROCBlock& RDOPROCTransact::getBlock()
+{
+	return m_block;
+}
+
+void RDOPROCTransact::setBlock(const LPIPROCBlock& block)
+{
+	m_block = block;
+}
+
 LPRDOResource RDOPROCTransact::clone(const LPRDORuntime& pRuntime) const
 {
 	LPRDOResource pResource = rdo::Factory<RDOPROCTransact>::create(pRuntime, getParamList(), getResType(), getTraceID(), getType(), traceable(), m_temporary);
@@ -138,52 +202,16 @@ RDOPROCResource::RDOPROCResource(const LPRDORuntime& pRuntime, const std::vector
 RDOPROCResource::~RDOPROCResource()
 {}
 
+std::string RDOPROCResource::whoAreYou()
+{
+	return "procRes";
+}
+
 LPRDOResource RDOPROCResource::clone(const LPRDORuntime& pRuntime) const
 {
 	LPRDOResource pResource = rdo::Factory<RDOPROCResource>::create(pRuntime, getParamList(), getResType(), getTraceID(), getType(), traceable(), m_temporary);
 	ASSERT(pResource);
 	return pResource;
-}
-
-// --------------------------------------------------------------------------------
-// -------------------- RDOPROCBlock
-// --------------------------------------------------------------------------------
-RDOPROCBlock::RDOPROCBlock(LPIPROCProcess pProcess)
-	: m_process(pProcess)
-{}
-
-bool RDOPROCBlock::init()
-{
-	if (!m_process)
-		return false;
-
-	m_process.query_cast<IBaseOperationContainer>()->append(this);
-	return true;
-}
-
-RDOPROCBlock::TransactIt RDOPROCBlock::transactFind(const LPTransact& pTransact)
-{
-	return std::find(m_transacts.begin(), m_transacts.end(), pTransact);
-}
-
-RDOPROCBlock::TransactIt RDOPROCBlock::transactEnd()
-{
-	return m_transacts.end();
-}
-
-void RDOPROCBlock::transactGoIn(const LPTransact& pTransact)
-{
-	m_transacts.push_back(pTransact);
-}
-
-void RDOPROCBlock::transactGoOut(const LPTransact& pTransact)
-{
-	m_transacts.remove(pTransact);
-}
-
-LPIPROCProcess RDOPROCBlock::getProcess() const
-{
-	return m_process;
 }
 
 CLOSE_RDO_RUNTIME_NAMESPACE
