+++ simulator/runtime/process/seize_release.h
@@ -112,6 +112,4 @@ private:
 
 CLOSE_RDO_RUNTIME_NAMESPACE
 
-#include "simulator/runtime/process/seize_release-inl.h"
-
 #endif // _LIB_RUNTIME_PROCESS_SEIZE_RELEASE_H_
