+++ simulator/runtime/process/terminate.h
@@ -50,6 +50,4 @@ private:
 
 CLOSE_RDO_RUNTIME_NAMESPACE
 
-#include "simulator/runtime/process/terminate-inl.h"
-
 #endif // _LIB_RUNTIME_PROCESS_TERMINATE_H_
